Get adding indicator rows and subindicator rows working
Including ensuring there is always at least one subindicator row, both
on page load, and when adding a subindicator row.

Along the way we've refactored a bit of code, come up with conventions
for how the indexes of the indicator and subindicator rows are encoded
into form element ids and names.